Refactoring DTO's & fix tests
