--- AI-Social-Platform.Services.Data.Models/PublicationDtos/LikeDto.cs
@@ -1,6 +1,0 @@
-﻿namespace AI_Social_Platform.Services.Data.Models.PublicationDtos
-{
-    public class LikeDto : BaseSocialDto
-    {
-    }
-}
