--- AI-Social-Platform.Services.Data.Models/PublicationDtos/ShareDto.cs
@@ -1,6 +1,0 @@
-﻿namespace AI_Social_Platform.Services.Data.Models.PublicationDtos
-{
-    public class ShareDto : BaseSocialDto
-    {
-    }
-}
