+++ AI-Social-Platform.Services.Data.Models/SocialFeature/BaseSocialDto.cs
@@ -1,4 +1,4 @@
-﻿namespace AI_Social_Platform.Services.Data.Models.PublicationDtos
+﻿namespace AI_Social_Platform.Services.Data.Models.SocialFeature
 {
     public class BaseSocialDto
     {
