+++ AI-Social-Platform.Services.Data.Models/SocialFeature/CommentDto.cs
@@ -1,7 +1,7 @@
-﻿namespace AI_Social_Platform.Services.Data.Models.PublicationDtos
+﻿namespace AI_Social_Platform.Services.Data.Models.SocialFeature
 {
     public class CommentDto : BaseSocialDto
-    { 
+    {
         public string Content { get; set; } = null!;
     }
 }
