+++ AI-Social-Platform.Services.Data.Models/SocialFeature/CommentFormDto.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using static AI_Social_Platform.Common.EntityValidationConstants.Comment;
 
-namespace AI_Social_Platform.Services.Data.Models.PublicationDtos
+namespace AI_Social_Platform.Services.Data.Models.SocialFeature
 {
     public class CommentFormDto
     {
@@ -9,6 +9,6 @@ namespace AI_Social_Platform.Services.Data.Models.PublicationDtos
         [StringLength(CommentContentMaxLength, MinimumLength = 1)]
         public string Content { get; set; } = null!;
 
-        public Guid PublicationId { get; set; } 
+        public Guid PublicationId { get; set; }
     }
 }
