+++ AI-Social-Platform.Services.Data.Models/SocialFeature/LikeDto.cs
@@ -1,0 +1,6 @@
+﻿namespace AI_Social_Platform.Services.Data.Models.SocialFeature
+{
+    public class LikeDto : BaseSocialDto
+    {
+    }
+}
