+++ Ai-Social-Platform.Tests/PublicationServiceTest.cs
@@ -2,12 +2,14 @@
 using AI_Social_Platform.Data.Models.Publication;
 using AI_Social_Platform.Services.Data;
 using AI_Social_Platform.Services.Data.Models.PublicationDtos;
-
+using AI_Social_Platform.Services.Data.Models.SocialFeature;
 using AutoMapper;
 
 using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using System.Security.Claims;
+using AI_Social_Platform.Data.Models;
+using AI_Social_Platform.Services.Data.Models.UserDto;
 using static AI_Social_Platform.Common.ExceptionMessages.PublicationExceptionMessages;
 
 namespace Ai_Social_Platform.Tests
@@ -61,6 +63,7 @@ namespace Ai_Social_Platform.Tests
                 x.CreateMap<ShareDto, Share>().ReverseMap();
                 x.CreateMap<PublicationDto, Publication>().ReverseMap();
                 x.CreateMap<PublicationFormDto, Publication>().ReverseMap();
+                x.CreateMap<UserDto, ApplicationUser>().ReverseMap();
             });
             mapper = config.CreateMapper();
             baseSocialService = new BaseSocialService(dbContext, httpContextAccessor, mapper);
@@ -81,10 +84,10 @@ namespace Ai_Social_Platform.Tests
 
             // Assert
             Assert.IsNotNull(result);
-            Assert.IsInstanceOf<IEnumerable<PublicationDto>>(result);
+            Assert.IsInstanceOf<IndexPublicationDto>(result);
 
             // Assuming some publications were seeded during setup
-            Assert.IsTrue(result.Any());
+            Assert.IsTrue(result.Publications.Any());
         }
 
         [Test]
@@ -322,7 +325,7 @@ namespace Ai_Social_Platform.Tests
             var countBefore = dbContext.Comments.Count(c => c.PublicationId == dto.PublicationId);
 
             // Act
-            await publicationService.CreateCommentAsync(dto);
+            await baseSocialService.CreateCommentAsync(dto);
 
             // Assert
             Assert.That(dbContext.Comments.Count(c => c.PublicationId == dto.PublicationId),
@@ -340,7 +343,7 @@ namespace Ai_Social_Platform.Tests
             };
 
             // Act
-            await publicationService.CreateCommentAsync(dto);
+            await baseSocialService.CreateCommentAsync(dto);
 
             // Assert
             Assert.That(dbContext.Comments.Last().Content, Is.EqualTo(dto.Content));
@@ -359,7 +362,7 @@ namespace Ai_Social_Platform.Tests
             // Act and Assert
             var exception = Assert.ThrowsAsync<NullReferenceException>(async () =>
             {
-                await publicationService.CreateCommentAsync(dto);
+                await baseSocialService.CreateCommentAsync(dto);
             });
 
             // Optionally assert on the exception message or other details
@@ -378,7 +381,7 @@ namespace Ai_Social_Platform.Tests
                 dbContext.Comments.First(c => c.UserId.ToString() == "123400ce-d726-4fc8-83d9-d6b3ac1f591e");
 
             // Act
-            await publicationService.UpdateCommentAsync(dto, comment.Id);
+            await baseSocialService.UpdateCommentAsync(dto, comment.Id);
 
             // Assert
             Assert.That(comment.Content, Is.EqualTo(dto.Content));
@@ -397,7 +400,7 @@ namespace Ai_Social_Platform.Tests
             // Act and Assert
             var exception = Assert.ThrowsAsync<NullReferenceException>(async () =>
             {
-                await publicationService.UpdateCommentAsync(dto, invalidCommentId);
+                await baseSocialService.UpdateCommentAsync(dto, invalidCommentId);
             });
 
             // Optionally assert on the exception message or other details
@@ -418,7 +421,7 @@ namespace Ai_Social_Platform.Tests
             // Act and Assert
             var exception = Assert.ThrowsAsync<AccessViolationException>(async () =>
             {
-                await publicationService.UpdateCommentAsync(dto, comment.Id);
+                await baseSocialService.UpdateCommentAsync(dto, comment.Id);
             });
 
             // Optionally assert on the exception message or other details
@@ -434,7 +437,7 @@ namespace Ai_Social_Platform.Tests
             var countBefore = dbContext.Comments.Count();
 
             // Act
-            await publicationService.DeleteCommentAsync(comment.Id);
+            await baseSocialService.DeleteCommentAsync(comment.Id);
 
             // Assert
             Assert.That(dbContext.Comments.Count(), Is.EqualTo(countBefore - 1));
@@ -449,7 +452,7 @@ namespace Ai_Social_Platform.Tests
             // Act and Assert
             var exception = Assert.ThrowsAsync<NullReferenceException>(async () =>
             {
-                await publicationService.DeleteCommentAsync(invalidCommentId);
+                await baseSocialService.DeleteCommentAsync(invalidCommentId);
             });
 
             // Optionally assert on the exception message or other details
@@ -466,7 +469,7 @@ namespace Ai_Social_Platform.Tests
             // Act and Assert
             var exception = Assert.ThrowsAsync<AccessViolationException>(async () =>
             {
-                await publicationService.DeleteCommentAsync(comment.Id);
+                await baseSocialService.DeleteCommentAsync(comment.Id);
             });
 
             // Optionally assert on the exception message or other details
@@ -480,7 +483,7 @@ namespace Ai_Social_Platform.Tests
             var publicationId = Guid.Parse("a0a0a6a0-0b1e-4b9e-9b0a-0b9b9b9b9b9b");
 
             // Act
-            var result = await publicationService.GetCommentsOnPublicationAsync(publicationId);
+            var result = await baseSocialService.GetCommentsOnPublicationAsync(publicationId);
 
             // Assert
             Assert.IsNotNull(result);
@@ -499,7 +502,7 @@ namespace Ai_Social_Platform.Tests
             var countBefore = dbContext.Likes.Count(l => l.PublicationId == publicationId);
 
             // Act
-            await publicationService.CreateLikesOnPublicationAsync(publicationId);
+            await baseSocialService.CreateLikesOnPublicationAsync(publicationId);
 
             // Assert
             Assert.That(dbContext.Likes.Count(l => l.PublicationId == publicationId), Is.EqualTo(countBefore + 1));
@@ -514,7 +517,7 @@ namespace Ai_Social_Platform.Tests
             // Act and Assert
             var exception = Assert.ThrowsAsync<NullReferenceException>(async () =>
             {
-                await publicationService.CreateLikesOnPublicationAsync(invalidPublicationId);
+                await baseSocialService.CreateLikesOnPublicationAsync(invalidPublicationId);
             });
 
             // Optionally assert on the exception message or other details
@@ -530,7 +533,7 @@ namespace Ai_Social_Platform.Tests
             var countBefore = dbContext.Likes.Count(l => l.PublicationId == publicationId);
 
             // Act
-            await publicationService.CreateLikesOnPublicationAsync(publicationId);
+            await baseSocialService.CreateLikesOnPublicationAsync(publicationId);
 
             // Assert
             Assert.That(dbContext.Likes.Count(l => l.PublicationId == publicationId), Is.EqualTo(countBefore + 1));
@@ -545,7 +548,7 @@ namespace Ai_Social_Platform.Tests
             var countBefore = dbContext.Likes.Count();
 
             // Act
-            await publicationService.DeleteLikeOnPublicationAsync(likeId);
+            await baseSocialService.DeleteLikeOnPublicationAsync(likeId);
 
             // Assert
             Assert.That(dbContext.Likes.Count(), Is.EqualTo(countBefore - 1));
@@ -560,7 +563,7 @@ namespace Ai_Social_Platform.Tests
             // Act and Assert
             var exception = Assert.ThrowsAsync<NullReferenceException>(async () =>
             {
-                await publicationService.DeleteLikeOnPublicationAsync(invalidLikeId);
+                await baseSocialService.DeleteLikeOnPublicationAsync(invalidLikeId);
             });
 
             // Optionally assert on the exception message or other details
@@ -576,7 +579,7 @@ namespace Ai_Social_Platform.Tests
             // Act and Assert
             var exception = Assert.ThrowsAsync<AccessViolationException>(async () =>
             {
-                await publicationService.DeleteLikeOnPublicationAsync(likeId);
+                await baseSocialService.DeleteLikeOnPublicationAsync(likeId);
             });
 
             // Optionally assert on the exception message or other details
@@ -590,7 +593,7 @@ namespace Ai_Social_Platform.Tests
             var publicationId = Guid.Parse("a0a0a6a0-0b1e-4b9e-9b0a-0b9b9b9b9b9b");
 
             // Act
-            var result = await publicationService.GetLikesOnPublicationAsync(publicationId);
+            var result = await baseSocialService.GetLikesOnPublicationAsync(publicationId);
 
             // Assert
             Assert.IsNotNull(result);
@@ -609,7 +612,7 @@ namespace Ai_Social_Platform.Tests
             var countBefore = dbContext.Shares.Count(l => l.PublicationId == publicationId);
 
             // Act
-            await publicationService.CreateSharesOnPublicationAsync(publicationId);
+            await baseSocialService.CreateSharesOnPublicationAsync(publicationId);
 
             // Assert
             Assert.That(dbContext.Shares.Count(l => l.PublicationId == publicationId), Is.EqualTo(countBefore + 1));
@@ -624,7 +627,7 @@ namespace Ai_Social_Platform.Tests
             // Act and Assert
             var exception = Assert.ThrowsAsync<NullReferenceException>(async () =>
             {
-                await publicationService.CreateSharesOnPublicationAsync(invalidPublicationId);
+                await baseSocialService.CreateSharesOnPublicationAsync(invalidPublicationId);
             });
 
             // Optionally assert on the exception message or other details
@@ -640,7 +643,7 @@ namespace Ai_Social_Platform.Tests
             var countBefore = dbContext.Shares.Count(l => l.PublicationId == publicationId);
 
             // Act
-            await publicationService.CreateSharesOnPublicationAsync(publicationId);
+            await baseSocialService.CreateSharesOnPublicationAsync(publicationId);
 
             // Assert
             Assert.That(dbContext.Shares.Count(l => l.PublicationId == publicationId), Is.EqualTo(countBefore + 1));
@@ -655,7 +658,7 @@ namespace Ai_Social_Platform.Tests
             var countBefore = dbContext.Shares.Count();
 
             // Act
-            await publicationService.DeleteShareOnPublicationAsync(shareId);
+            await baseSocialService.DeleteShareOnPublicationAsync(shareId);
 
             // Assert
             Assert.That(dbContext.Shares.Count(), Is.EqualTo(countBefore - 1));
@@ -670,7 +673,7 @@ namespace Ai_Social_Platform.Tests
             // Act and Assert
             var exception = Assert.ThrowsAsync<NullReferenceException>(async () =>
             {
-                await publicationService.DeleteShareOnPublicationAsync(invalidShareId);
+                await baseSocialService.DeleteShareOnPublicationAsync(invalidShareId);
             });
 
             // Optionally assert on the exception message or other details
@@ -686,7 +689,7 @@ namespace Ai_Social_Platform.Tests
             // Act and Assert
             var exception = Assert.ThrowsAsync<AccessViolationException>(async () =>
             {
-                await publicationService.DeleteShareOnPublicationAsync(shareId);
+                await baseSocialService.DeleteShareOnPublicationAsync(shareId);
             });
 
             // Optionally assert on the exception message or other details
@@ -700,7 +703,7 @@ namespace Ai_Social_Platform.Tests
             var publicationId = Guid.Parse("a0a0a6a0-0b1e-4b9e-9b0a-0b9b9b9b9b9b");
 
             // Act
-            var result = await publicationService.GetSharesOnPublicationAsync(publicationId);
+            var result = await baseSocialService.GetSharesOnPublicationAsync(publicationId);
 
             // Assert
             Assert.IsNotNull(result);
